`experimental mask_user_info`を`filesystem mask_user_info`に移動する
