+++ annofabcli/filesystem/mask_user_info.py
