+++ annofabcli/filesystem/subcommand_filesystem.py
@@ -3,6 +3,7 @@ import argparse
 import annofabcli
 import annofabcli.common.cli
 import annofabcli.filesystem.filter_annotation
+import annofabcli.filesystem.mask_user_info
 import annofabcli.filesystem.write_annotation_image
 
 
@@ -12,6 +13,7 @@ def parse_args(parser: argparse.ArgumentParser):
 
     # サブコマンドの定義
     annofabcli.filesystem.filter_annotation.add_parser(subparsers)
+    annofabcli.filesystem.mask_user_info.add_parser(subparsers)
     annofabcli.filesystem.write_annotation_image.add_parser(subparsers)
