+++ docs/command_reference/labor/index.rst
@@ -15,8 +15,5 @@ Available Commands
    :maxdepth: 1
    :titlesonly:
 
-   delete
-   list
-   list_last
-   list_task_creation_history
-   wait
+   list_worktime_by_user
+
