--- tests/data/experimental/.~lock.user2.csv#
@@ -1,1 +1,0 @@
-,vagrant,example,07.08.2020 18:20,file:///home/vagrant/.config/libreoffice/4;
