+++ tests/data/filesystem/label_color.json
@@ -1,0 +1,37 @@
+{
+  "Cat": [
+    255,
+    99,
+    71
+  ],
+  "leg": [
+    27,
+    144,
+    185
+  ],
+  "eye": [
+    88,
+    113,
+    249
+  ],
+  "dog": [
+    188,
+    83,
+    41
+  ],
+  "human": [
+    210,
+    54,
+    28
+  ],
+  "bird": [
+    29,
+    202,
+    101
+  ],
+  "climatic": [
+    255,
+    255,
+    255
+  ]
+}
